+++ examples/with-context-api/pages/about.tsx
@@ -1,0 +1,34 @@
+import type { MouseEvent } from 'react'
+import Link from 'next/link'
+import { useCount, useDispatchCount } from '../components/Counter'
+
+const AboutPage = () => {
+  const count = useCount()
+  const dispatch = useDispatchCount()
+
+  const handleIncrease = (event: MouseEvent<HTMLButtonElement>) =>
+    dispatch({
+      type: 'INCREASE',
+    })
+  const handleIncrease15 = (event: MouseEvent<HTMLButtonElement>) =>
+    dispatch({
+      type: 'INCREASE_BY',
+      payload: 15,
+    })
+
+  return (
+    <>
+      <h1>ABOUT</h1>
+      <p>Counter: {count}</p>
+      <button onClick={handleIncrease}>Increase</button>
+      <button onClick={handleIncrease15}>Increase By 15</button>
+      <p>
+        <Link href="/">
+          <a>Home</a>
+        </Link>
+      </p>
+    </>
+  )
+}
+
+export default AboutPage
