+++ examples/with-context-api/pages/index.tsx
@@ -1,33 +1,33 @@
+import type { MouseEvent } from 'react'
 import Link from 'next/link'
 import { useCount, useDispatchCount } from '../components/Counter'
 
-const AboutPage = () => {
+const IndexPage = () => {
   const count = useCount()
   const dispatch = useDispatchCount()
 
-  const handleIncrease = (event) =>
+  const handleIncrease = (event: MouseEvent<HTMLButtonElement>) =>
     dispatch({
       type: 'INCREASE',
     })
-  const handleIncrease15 = (event) =>
+  const handleDecrease = (event: MouseEvent<HTMLButtonElement>) =>
     dispatch({
-      type: 'INCREASE_BY',
-      payload: 15,
+      type: 'DECREASE',
     })
 
   return (
     <>
-      <h1>ABOUT</h1>
+      <h1>HOME</h1>
       <p>Counter: {count}</p>
       <button onClick={handleIncrease}>Increase</button>
-      <button onClick={handleIncrease15}>Increase By 15</button>
+      <button onClick={handleDecrease}>Decrease</button>
       <p>
-        <Link href="/">
-          <a>Home</a>
+        <Link href="/about">
+          <a>About</a>
         </Link>
       </p>
     </>
   )
 }
 
-export default AboutPage
+export default IndexPage
